Fix circular dependency without moving AnnotateDetectorsOnLayerNode
