+++ src/tqec/compile/tree/annotators/detectors.py
@@ -2,7 +2,14 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 
+from typing_extensions import override
+
 from tqec.circuit.measurement_map import MeasurementRecordsMap
+from tqec.compile.blocks.layers.atomic.layout import LayoutLayer
+from tqec.compile.detectors.compute import compute_detectors_for_fixed_radius
+from tqec.compile.detectors.database import DetectorDatabase
+from tqec.compile.tree.annotations import DetectorAnnotation
+from tqec.compile.tree.node import LayerNode, NodeWalker
 from tqec.plaquette.plaquette import Plaquettes
 from tqec.templates.base import Template
 from tqec.utils.exceptions import TQECError
@@ -149,3 +156,101 @@ class LookbackStack:
                 "Cannot get a meaningful stack length when a REPEAT block is in construction."
             )
         return len(self._stack[0])
+
+
+class AnnotateDetectorsOnLayerNode(NodeWalker):
+    def __init__(
+        self,
+        k: int,
+        manhattan_radius: int = 2,
+        detector_database: DetectorDatabase | None = None,
+        lookback: int = 2,
+        parallel_process_count: int = 1,
+    ):
+        """Walker computing and annotating detectors on leaf nodes.
+
+        This class keeps track of the ``lookback`` previous leaf nodes seen and
+        uses them to automatically compute the detectors at all the leaf nodes
+        it encounters.
+
+        Args:
+            k: scaling factor.
+            manhattan_radius: Parameter for the automatic computation of detectors.
+                Should be large enough so that flows cancelling each other to
+                form a detector are strictly contained in plaquettes that are at
+                most at a distance of ``manhattan_radius`` from the central
+                plaquette. Detector computation runtime grows with this parameter,
+                so you should try to keep it to its minimum. A value too low might
+                produce invalid detectors.
+            detector_database: existing database of detectors that is used to
+                avoid computing detectors if the database already contains them.
+                Default to `None` which result in not using any kind of database
+                and unconditionally performing the detector computation.
+                in the database is encountered. Default to ``False``.
+            lookback: number of QEC rounds to consider to try to find detectors. Including more
+                rounds increases computation time.
+            parallel_process_count: number of processes to use for parallel processing.
+                1 for sequential processing, >1 for parallel processing using
+                ``parallel_process_count`` processes, and -1 for using all available
+                CPU cores. Default to 1.
+
+        """
+        if lookback < 1:
+            raise TQECError(
+                "Cannot compute detectors without any layer. The `lookback` "
+                f"parameter should be >= 1 but got {lookback}."
+            )
+        self._k = k
+        self._manhattan_radius = manhattan_radius
+        self._database = detector_database if detector_database is not None else DetectorDatabase()
+        self._lookback_size = lookback
+        self._lookback_stack = LookbackStack()
+        self._parallel_process_count = parallel_process_count
+
+    @override
+    def visit_node(self, node: LayerNode) -> None:
+        if not isinstance(node._layer, LayoutLayer):
+            return
+        annotations = node.get_annotations(self._k)
+        if annotations.circuit is None:
+            raise TQECError("Cannot compute detectors without the circuit annotation.")
+        self._lookback_stack.append(
+            *node._layer.to_template_and_plaquettes(),
+            MeasurementRecordsMap.from_scheduled_circuit(annotations.circuit),
+        )
+        templates, plaquettes, measurement_records = self._lookback_stack.lookback(
+            self._lookback_size
+        )
+
+        detectors = compute_detectors_for_fixed_radius(
+            templates,
+            self._k,
+            plaquettes,
+            self._manhattan_radius,
+            self._database,
+            only_use_database=False,
+            parallel_process_count=self._parallel_process_count,
+        )
+
+        for detector in detectors:
+            annotations.detectors.append(
+                DetectorAnnotation.from_detector(detector, measurement_records)
+            )
+
+    @override
+    def enter_node(self, node: LayerNode) -> None:
+        if node.is_repeated:
+            self._lookback_stack.enter_repeat_block()
+
+    @override
+    def exit_node(self, node: LayerNode) -> None:
+        if not node.is_repeated:
+            return
+        # Note: this is the place to perform checks. In particular, checking that
+        # detectors computed at the first repetition of the REPEAT block are also
+        # valid at any repetitions. This is a requirement for the REPEAT block to
+        # make sense, but that would be nice to include a check to avoid
+        # misleadingly include detectors that are incorrect sometimes.
+        repetitions = node.repetitions
+        assert repetitions is not None
+        self._lookback_stack.close_repeat_block(repetitions.integer_eval(self._k))
